RAG + Langchain Chatbot/requirements.txt
langchain
python-dotenv
unstructured
markdown
langchain-google-genai
google-generativeai
langchain-community
langchain-core
faiss-cpu